requirements.txt
Flask==2.0.3
werkzeug==2.0.3
vonage==3.0.0
openai==0.27.0
python-dotenv==0.19.0
echo "gunicorn==20.1.0" >> requirements.txt
gunicorn==20.1.0
